Allow for replica only keepers
We want to be able to use replica only keepers that are never eligible
for master election or synchronous standby promotion.

We do this by adding a --replica flag to the keeper that is propagated to
the keeper spec and to the keepers db spec allowing us to filter out
these keepers/dbs during the initial master election and synchronous
standby promotion.